requirements.txt
streamlit>=1.32,<2.0
tensorflow==2.17.0
numpy>=1.26,<2.0
pillow>=10.0
pandas>=2.0
matplotlib>=3.8
